Fix offsets.

+++ Penumbra/Interop/Hooks/Meta/GetEqpIndirect.cs
@@ -22,13 +22,12 @@ public sealed unsafe class GetEqpIndirect : FastHook<GetEqpIndirect.Delegate>
     [MethodImpl(MethodImplOptions.AggressiveInlining | MethodImplOptions.AggressiveOptimization)]
     private void Detour(DrawObject* drawObject)
     {
-        if ((*(byte*)(drawObject + Offsets.GetEqpIndirectSkip1) & 1) == 0 || *(ulong*)(drawObject + Offsets.GetEqpIndirectSkip2) == 0)
+        // Shortcut because this is also called all the time.
+        // Same thing is checked at the beginning of the original function.
+        if ((*(byte*)((nint)drawObject + Offsets.GetEqpIndirectSkip1) & 1) == 0 || *(ulong*)((nint)drawObject + Offsets.GetEqpIndirectSkip2) == 0)
             return;
 
         Penumbra.Log.Excessive($"[Get EQP Indirect] Invoked on {(nint)drawObject:X}.");
-        // Shortcut because this is also called all the time.
-        // Same thing is checked at the beginning of the original function.
-
         var       collection = _collectionResolver.IdentifyCollection(drawObject, true);
         using var eqp        = _metaState.ResolveEqpData(collection.ModCollection);
         Task.Result.Original(drawObject);
