+++ Penumbra/Interop/Hooks/Meta/UpdateModel.cs
@@ -24,7 +24,7 @@ public sealed unsafe class UpdateModel : FastHook<UpdateModel.Delegate>
     {
         // Shortcut because this is called all the time.
         // Same thing is checked at the beginning of the original function.
-        if (*(int*)(drawObject + Offsets.UpdateModelSkip) == 0)
+        if (*(int*)((nint)drawObject + Offsets.UpdateModelSkip) == 0)
             return;
 
         Penumbra.Log.Excessive($"[Update Model] Invoked on {(nint)drawObject:X}.");
